backend/create-payment/requirements.txt
yookassa>=3.0.0
psycopg2-binary
